Add files via upload
The algorithm is very simple: at first it creates a bunc of (x,y) points and then it calculates z-values for every pont. After that it takes a point with a smallest value en declare it as "The Leader of the Pack". Then it pushes all other pack members (points) towards the leader, then it takes some of them and rotates them for chosen (or random) angle and finally it calulates brand new z-values... The algorithm is repeating itselves for a chosen number of iterations. No selection of a better chromosomes (pack members) for now...